requirements.txt
asn1crypto
certifi
cffi
cryptography
future
aio-pika
pika
pycparser
python-telegram-bot[job-queue]
redis
six
tornado
